requirements.txt
contourpy==1.1.0
cycler==0.11.0
drake==1.20.0
fonttools==4.42.1
kiwisolver==1.4.5
matplotlib==3.7.2
meshcat==0.3.2
numpy==1.25.2
packaging==23.1
Pillow==10.0.0
pydot==1.4.2
pyngrok==6.0.0
pyparsing==3.0.9
python-dateutil==2.8.2
PyYAML==6.0.1
six==1.16.0
u-msgpack-python==2.8.0
